Extract menu actions in base activity.
